+++ src/main/java/com/bitpay/sdk/model/Invoice/MinerFeesItem.java
@@ -4,32 +4,34 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigDecimal;
+
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class MinerFeesItem {
 
-    private Double _satoshisPerByte;
-    private Double _totalFee;
+    private BigDecimal _satoshisPerByte;
+    private BigDecimal _totalFee;
 
     public MinerFeesItem() {
     }
 
     @JsonIgnore
-    public Double getSatoshisPerByte() {
+    public BigDecimal getSatoshisPerByte() {
         return _satoshisPerByte;
     }
 
     @JsonProperty("satoshisPerByte")
-    public void setSatoshisPerByte(Double satoshisPerByte) {
+    public void setSatoshisPerByte(BigDecimal satoshisPerByte) {
         this._satoshisPerByte = satoshisPerByte;
     }
 
     @JsonIgnore
-    public Double getTotalFee() {
+    public BigDecimal getTotalFee() {
         return _totalFee;
     }
 
     @JsonProperty("totalFee")
-    public void setTotalFee(Double totalFee) {
+    public void setTotalFee(BigDecimal totalFee) {
         this._totalFee = totalFee;
     }
 }
